fix: grouping bug caused by select/ linked records
Linked records if they do not have the same interface would create inconsistent unique keys data for which would not be available and thus would create incorrect groups. This only affected the implementation when data changed.
The change makes the implementation opinionated and objects as values are only allowed if they have an id property to create unique ids